Refactored Sharepoint capability to move code out of BaseBrowserHandler.cs and into SharepointHelper.cs to reduce code size of BaseBrowserHandler.cs
