+++ src/Ghosts.Client/Ghosts.Client.csproj
@@ -543,6 +543,7 @@
     <Compile Include="Handlers\BaseBrowserHandler.cs" />
     <Compile Include="Handlers\BrowserCrawl.cs" />
     <Compile Include="Handlers\Print.cs" />
+    <Compile Include="Handlers\SharepointHelper.cs" />
     <Compile Include="Handlers\Ssh.cs" />
     <Compile Include="Infrastructure\Browser\ExtendedConfiguration.cs" />
     <Compile Include="Infrastructure\Credentials.cs" />
